maze stuff

+++ ChronoLoop/ChronoLoop/Actions/CCLoadHub.h
@@ -150,12 +150,12 @@ namespace Epoch
 							temp == "mmTutButton" || temp == "mmTutSign" || temp == "mmTutStand")
 						{
 							Transform t;
-							t.SetMatrix(((BaseObject*)*it)->GetTransform().GetMatrix() * matrix4::CreateTranslation(0, floorPos, 0));
+							t.SetMatrix(((BaseObject*)*it)->GetTransform().GetMatrix() * matrix4::CreateTranslation(0, (float)floorPos, 0));
 							((BaseObject*)*it)->SetTransform(t);
 
 							if(boop)
 							{
-								next->GetStartPos() = vec4f(0, floorPos, 0, 1);
+								next->GetStartPos() = vec4f(0, (float)floorPos, 0, 1);
 								boop = false;
 							}
 
@@ -203,7 +203,7 @@ namespace Epoch
 						p1->SetColors(vec3f(0, 1, 0), vec3f(0, .5f, .5f));
 						p1->SetLife(550);
 						p1->SetSize(.35f, .15f);
-						ParticleEmitter* emit11 = new TeleportEffect(-1, 150, 2, vec4f(0, -10, 2.820054, 1));
+						ParticleEmitter* emit11 = new TeleportEffect(-1, 150, 2, vec4f(0, -10, 2.820054f, 1));
 						emit11->SetParticle(p1);
 						emit11->SetTexture("../Resources/BasicRectP.png");
 						((TeleportEffect*)emit11)->y1 = 8;
@@ -218,7 +218,7 @@ namespace Epoch
 						p1->SetColors(vec3f(0, .5f, .5f), vec3f(0, 1, 0));
 						p1->SetLife(550);
 						p1->SetSize(.15f, .05f);
-						ParticleEmitter* emit12 = new TeleportEffect(-1, 150, 2, vec4f(0, -10, 2.820054, 1));
+						ParticleEmitter* emit12 = new TeleportEffect(-1, 150, 2, vec4f(0, -10, 2.820054f, 1));
 						emit12->SetTexture("../Resources/BasicCircleP.png");
 						emit12->SetParticle(p1);
 						((TeleportEffect*)emit12)->y1 = 1;
@@ -235,7 +235,7 @@ namespace Epoch
 						p1->SetColors(vec3f(1, 0, 0), vec3f(.5f, 0, .5f));
 						p1->SetLife(550);
 						p1->SetSize(.35f, .15f);
-						ParticleEmitter* emit11 = new TeleportEffect(-1, 150, 2, vec4f(0, -10, 2.820054, 1));
+						ParticleEmitter* emit11 = new TeleportEffect(-1, 150, 2, vec4f(0, -10, 2.820054f, 1));
 						emit11->SetParticle(p1);
 						emit11->SetTexture("../Resources/BasicRectP.png");
 						((TeleportEffect*)emit11)->y1 = 8;
@@ -250,7 +250,7 @@ namespace Epoch
 						p1->SetColors(vec3f(.5f, 0, .5f), vec3f(1, 0, 0));
 						p1->SetLife(550);
 						p1->SetSize(.15f, .05f);
-						ParticleEmitter* emit12 = new TeleportEffect(-1, 150, 2, vec4f(0, -10, 2.820054, 1));
+						ParticleEmitter* emit12 = new TeleportEffect(-1, 150, 2, vec4f(0, -10, 2.820054f, 1));
 						emit12->SetTexture("../Resources/BasicCircleP.png");
 						emit12->SetParticle(p1);
 						((TeleportEffect*)emit12)->y1 = 1;
@@ -268,7 +268,7 @@ namespace Epoch
 						p2->SetColors(vec3f(0, 1, 0), vec3f(0, .5f, .5f));
 						p2->SetLife(500);
 						p2->SetSize(.35f, .15f);
-						ParticleEmitter* emit21 = new TeleportEffect(-1, 150, 2, vec4f(-2.82, -10, 0, 1));
+						ParticleEmitter* emit21 = new TeleportEffect(-1, 150, 2, vec4f(-2.82f, -10, 0, 1));
 						emit21->SetParticle(p2);
 						emit21->SetTexture("../Resources/BasicRectP.png");
 						((TeleportEffect*)emit21)->y1 = 8;
@@ -283,7 +283,7 @@ namespace Epoch
 						p2->SetColors(vec3f(0, .5f, .5f), vec3f(0, 1, 0));
 						p2->SetLife(500);
 						p2->SetSize(.15f, .05f);
-						ParticleEmitter* emit22 = new TeleportEffect(-1, 150, 2, vec4f(-2.82, -10, 0, 1));
+						ParticleEmitter* emit22 = new TeleportEffect(-1, 150, 2, vec4f(-2.82f, -10, 0, 1));
 						emit22->SetTexture("../Resources/BasicCircleP.png");
 						emit22->SetParticle(p2);
 						((TeleportEffect*)emit22)->y1 = 1;
@@ -300,7 +300,7 @@ namespace Epoch
 						p2->SetColors(vec3f(1, 0, 0), vec3f(.5f, 0, .5f));
 						p2->SetLife(500);
 						p2->SetSize(.35f, .15f);
-						ParticleEmitter* emit21 = new TeleportEffect(-1, 150, 2, vec4f(-2.82, -10, 0, 1));
+						ParticleEmitter* emit21 = new TeleportEffect(-1, 150, 2, vec4f(-2.82f, -10, 0, 1));
 						emit21->SetParticle(p2);
 						emit21->SetTexture("../Resources/BasicRectP.png");
 						((TeleportEffect*)emit21)->y1 = 8;
